Split files + clang-format
