+++ src/core/core_headers.h
@@ -114,7 +114,9 @@ class StackDump : public wxStackWalker // so we can give backtraces..
 #include "socket_codes.h"
 #include "template_matching.h"
 #include "functions.h"
-#include "run_profiles.h"
+#include "run_command.h"
+#include "run_profile.h"
+#include "run_profile_manager.h"
 #include "job_packager.h"
 #include "ctf.h"
 #include "curve.h"
