+++ src/gui/MyMovieAlignResultsPanel.h
@@ -42,5 +42,4 @@ class MyMovieAlignResultsPanel : public MovieAlignResultsPanel {
     bool group_combo_is_dirty;
 
     wxString current_fill_command;
-
 };
